Refactor systems to inherit from `BaseSystem` and streamline dynamics resolution into base class
